+++ CHANGELOG.rst
@@ -8,9 +8,14 @@ Development Version
 * Introduce ``LiteTranslator`` and ``FullTranslator`` as
   convenience classes for more deterministic behaviour, :pr:`95`.
 * ``setuptools`` plugin:
+
   * Fix ``ValueError`` when ``setup.cfg`` contains ``[options.packages.find]``
     but also lists ``[options] packages = ...`` explicitly as a list of package
-    names, :issue:`93`
+    names, :issue:`93`.
+
+* ``pytest`` plugin:
+
+  * Remove comments when converting ``addopts`` with multi-line values, :issue:`98`.
 
 Version 0.13
 ============